Added support for slice-parameters to be passed in as spans and read-only spans, instead of arrays.
The hope being this provides support for more exotic memory structures.

This is not behind a version check because it's an explicit option, and it will panic if span is selected but unsafe code support is not.

I'm not sure a good way to require this without a larger API change, and my goal was to make this a small opt-in feature.