Don't allow non-ignored service methods to take an incompatible receiver
Previously, services were allowed to have a service method that
had as its first argument a non-receiver (or ref to itself).

This would be the equivalent to a 'static' method generally, but is
currently not treated as such.

However, backends do expect that each 'service method' is actually...
using the method. As such, they could be given non-sensically
information and generate non-working code with that.

Signed-off-by: Marcel Müller <neikos@neikos.email>