chore(release): Fix dev-dependencies
As per https://github.com/rust-lang/cargo/issues/4242
and from cargo-workspaces and cargo publish docs:

Note: `dev-dependencies` are not taken into account when building
the dependency graph used to determine the proper publishing order.

This is because dev-dependencies are ignored by cargo publish - as
such, a dev-dependency on a local crate (with a path attribute),
should not have a version field.